Add files via upload
Sentiment analysis is the process of locating and categorizing the emotions represented in a text source. When analyzed, tweets can produce a significant amount of sentiment data. These statistics help us understand how individuals feel about a range of issues. Therefore, in order to calculate the consumer perspective, we need to create an automated machine learning sentiment analysis model